+++ test/dummy/app/controllers/sign_ups_controller.rb
@@ -1,8 +1,6 @@
 # To test sign-ups with confirmation (activation / verification)
 class SignUpsController < ApplicationController
 
-  around_action :toggle_confirmation
-
   def new
     @user = User.new
   end
@@ -11,9 +9,9 @@ class SignUpsController < ApplicationController
   def create
     @user = User.new user_params
     if @user.save
+      login @user
       if QuoVadis.accounts_require_confirmation
-        request_confirmation @user
-        redirect_to quo_vadis.confirmations_path
+        redirect_to secret_articles_path
       else
         redirect_to articles_path
       end
@@ -35,13 +33,6 @@ class SignUpsController < ApplicationController
 
   def user_params
     params.require(:user).permit(:name, :email, :password, :password_confirmation)
-  end
-
-  def toggle_confirmation
-    QuoVadis.accounts_require_confirmation true
-    yield
-  ensure
-    QuoVadis.accounts_require_confirmation false
   end
 
 end